use templates for operator and function definitions for native vector types.

+++ nrtdsm/gpu_kernels/nrtdsm_preprocess_kernels.cu
@@ -45,7 +45,7 @@ CUDA_DEVICE_KERNEL void generateMinMaxMipMap(
     const int2 dstPixIdx(
         blockDim.x * blockIdx.x + threadIdx.x,
         blockDim.y * blockIdx.y + threadIdx.y);
-    const int2 srcImageSize = material->heightMapSize >> srcMipLevel;
+    const int2 srcImageSize = material->heightMapSize >> static_cast<int32_t>(srcMipLevel);
     const int2 dstImageSize = srcImageSize / 2;
     if (dstPixIdx.x >= dstImageSize.x || dstPixIdx.y >= dstImageSize.y)
         return;
